init RL refactor PR
