+++ examples/four_boxes_example.py
@@ -49,7 +49,7 @@ def create_scene(port=55000):
         material=sm.Material.WHITE,
         physics_component=sm.RigidBodyComponent(mass=mass),
     )
-    scene += sm.SimpleRlAgent(
+    scene += sm.EgocentricCameraAgent(
         name="agent",
         sensors=[
             sm.CameraSensor(width=CAMERA_WIDTH, height=CAMERA_HEIGHT, position=[0, 0.1, 0]),
