+++ src/simenv/rl/__init__.py
@@ -1,5 +1,5 @@
 from .actions import *
+from .core import *
 from .reward_functions import *
 from .rl_agents import *
-from .rl_component import *
 from .wrappers import *
